refactor: slug page into simplier components and hooks
